Revert "Move editorOverrideService to common"
This reverts commit 84809f41cd7ad9842b397fd269188bf9033a3c15.
