+++ src/vs/workbench/contrib/preferences/browser/preferencesEditor.ts
@@ -38,8 +38,7 @@ import { scrollbarShadow } from 'vs/platform/theme/common/colorRegistry';
 import { attachStylerCallback } from 'vs/platform/theme/common/styler';
 import { IThemeService } from 'vs/platform/theme/common/themeService';
 import { IWorkspaceContextService } from 'vs/platform/workspace/common/workspace';
-import { IEditorRegistry } from 'vs/workbench/browser/editor';
-import { Extensions as EditorExtensions } from 'vs/workbench/services/editor/common/editorOverrideService';
+import { Extensions as EditorExtensions, IEditorRegistry } from 'vs/workbench/browser/editor';
 import { EditorPane } from 'vs/workbench/browser/parts/editor/editorPane';
 import { BaseTextEditor } from 'vs/workbench/browser/parts/editor/textEditor';
 import { EditorInput, EditorOptions, IEditorControl, IEditorOpenContext } from 'vs/workbench/common/editor';
