+++ src/vs/workbench/contrib/webviewPanel/browser/webviewPanel.contribution.ts
@@ -9,7 +9,7 @@ import { EditorOverride, ITextEditorOptions } from 'vs/platform/editor/common/ed
 import { SyncDescriptor } from 'vs/platform/instantiation/common/descriptors';
 import { registerSingleton } from 'vs/platform/instantiation/common/extensions';
 import { Registry } from 'vs/platform/registry/common/platform';
-import { EditorDescriptor, IEditorRegistry } from 'vs/workbench/browser/editor';
+import { EditorDescriptor, Extensions as EditorExtensions, IEditorRegistry } from 'vs/workbench/browser/editor';
 import { Extensions as WorkbenchExtensions, IWorkbenchContribution, IWorkbenchContributionsRegistry } from 'vs/workbench/common/contributions';
 import { Extensions as EditorInputExtensions, IEditorInput, IEditorInputFactoryRegistry } from 'vs/workbench/common/editor';
 import { IEditorGroup, IEditorGroupsService } from 'vs/workbench/services/editor/common/editorGroupsService';
@@ -20,8 +20,6 @@ import { WebviewEditor } from './webviewEditor';
 import { WebviewInput } from './webviewEditorInput';
 import { WebviewEditorInputSerializer } from './webviewEditorInputSerializer';
 import { IWebviewWorkbenchService, WebviewEditorService } from './webviewWorkbenchService';
-import { Extensions as EditorExtensions } from 'vs/workbench/services/editor/common/editorOverrideService';
-
 
 (Registry.as<IEditorRegistry>(EditorExtensions.Editors)).registerEditor(EditorDescriptor.create(
 	WebviewEditor,
